large_img_prediction/requirements.txt
tensorflow-gpu==1.7
keras==2.2.0
scikit-image
openslide-python
numpy==1.14.3
opencv-python
scipy
Pillow
cython
matplotlib
scikit-image
h5py
imgaug
ipython
lxml
progressbar
interval
descartes
shapely
